backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.34.0
websockets==13.1
google-cloud-aiplatform==1.87.0
google-adk==1.27.1
deprecated>=1.2.14
google-genai==1.12.1
google-cloud-firestore==2.19.0
google-cloud-pubsub==2.26.1
python-dotenv==1.0.1
httpx==0.28.1
elevenlabs==1.9.0
pydantic==2.9.2
python-jose[cryptography]==3.3.0
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-mock==3.14.0
